fix: 修复主流程画布功能

+++ src/pages/gundam-pages/main-draw/index.tsx
@@ -93,6 +93,12 @@ const MainDraw = (props: any) => {
       params.copyId = node.copy_id;
     }
 
+    if (node._nodetype === 'start') {
+      (fake.current as any).executeCommand?.('undo');
+      message.warning('开始节点不允许复制粘贴');
+      return;
+    }
+
     let res = await addNode(params);
     if (res === false) {
       (fake.current as any).deleteNode(node);
